+++ src/CubicAIExplorer/Models/DetailsColumnId.cs
@@ -1,0 +1,9 @@
+namespace CubicAIExplorer.Models;
+
+public enum DetailsColumnId
+{
+    Name,
+    Size,
+    Type,
+    DateModified
+}